fixed bug in previous bug fix; you can load workspaces and samples again in the slice viewer
Former-commit-id: cb146683c85125aa37a3f55fb91042aff6a1eceb

+++ console/src/main/java/org/janelia/it/FlyWorkstation/gui/viewer3d/slice_viewer/SliceViewViewer.java
@@ -61,7 +61,7 @@ public class SliceViewViewer extends Viewer {
     public void loadEntity(RootedEntity rootedEntity) {
         // don't reload if user tries to reload the same entity (is that a
         //  good idea?  not clear)
-        if (rootedEntity.getEntity().getId() != initialEntity.getId()) {
+        if (initialEntity != null && rootedEntity.getEntity().getId() != initialEntity.getId()) {
             deleteAll();
         }
         initialEntity = rootedEntity.getEntity();
